js/xpconnect/tests/chrome/test_chrometoSource.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<window title="Mozilla Bug 761723"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=761723" target="_blank">Mozilla Bug 761723</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript" src="outoflinexulscript.js"></script>
  <script type="application/javascript"><![CDATA[
const Cu = Components.utils;
Cu.import("resource://gre/modules/NetUtil.jsm");

function inlinefunction() {
         return 42;
}

isnot(inlinefunction.toSource().indexOf("return 42"), -1, "inline XUL script should have source");
isnot(outoflinefunction.toSource().indexOf("return 42"), -1, "outofline XUL script should have source");
isnot(NetUtil.asyncFetch.toSource().indexOf("return"), -1, "JSM module should have source");
  ]]></script>
</window>
